CMS Project/requirements.txt
Package    Version
---------- -------
Pillow     9.5.0
pip        23.1.2
setuptools 65.5.0